Move file service storage tests to their original package
