+++ pass-core-main/src/test/java/org/eclipse/pass/file/service/storage/FileStorageServiceS3Test.java
@@ -1,4 +1,4 @@
-package org.eclipse.pass.file.service;
+package org.eclipse.pass.file.service.storage;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertFalse;
@@ -11,11 +11,7 @@ import java.util.HashMap;
 import java.util.Map;
 
 import edu.wisc.library.ocfl.api.exception.NotFoundException;
-import org.eclipse.pass.file.service.storage.FileStorageService;
-import org.eclipse.pass.file.service.storage.StorageConfiguration;
-import org.eclipse.pass.file.service.storage.StorageFile;
-import org.eclipse.pass.file.service.storage.StorageProperties;
-import org.eclipse.pass.file.service.storage.StorageServiceType;
+import io.findify.s3mock.S3Mock;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
@@ -24,48 +20,51 @@ import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.util.FileSystemUtils;
 
-public class FileStorageServiceTest {
-    StorageConfiguration storageConfiguration;
+class FileStorageServiceS3Test {
+    private final static String ROOT_DIR = System.getProperty("java.io.tmpdir") + "/pass-s3-test";
+
     private FileStorageService fileStorageService;
     private final StorageProperties properties = new StorageProperties();
-    private final String rootDir = System.getProperty("java.io.tmpdir") + "/pass-file-system-test";
+    private S3Mock s3MockApi;
 
     /**
-     * Setup the FileStorageService for testing. Uses the system temp directory for the root directory.
+     * Setup the test environment. Uses custom endpoint for the in-memory S3 mock.
      */
     @BeforeEach
     void setUp() {
-        properties.setStorageType(StorageServiceType.FILE_SYSTEM);
-        properties.setRootDir(rootDir);
-        storageConfiguration =  new StorageConfiguration(properties);
+        s3MockApi = new S3Mock.Builder().withPort(8001).withInMemoryBackend().build();
+        s3MockApi.start();
+        properties.setStorageType(StorageServiceType.S3);
+        properties.setRootDir(ROOT_DIR);
+        properties.setS3Endpoint("http://localhost:8001");
+        properties.setS3BucketName("bucket-test-name");
+        properties.setS3RepoPrefix("s3-repo-prefix");
+        StorageConfiguration storageConfiguration = new StorageConfiguration(properties);
+
         try {
-            fileStorageService = new FileStorageService(storageConfiguration);
+            fileStorageService = new FileStorageService(storageConfiguration, "us-east-1");
         } catch (IOException e) {
             assertEquals("Exception during setup", e.getMessage());
         }
     }
 
     /**
-     * Cleanup the FileStorageService after testing. Deletes the root directory.
+     * Tear down the test environment. Deletes the temporary directory.
      */
     @AfterEach
     void tearDown() {
-        try {
-            FileSystemUtils.deleteRecursively(Paths.get(rootDir));
-        } catch (IOException e) {
-            assertEquals("An exception was thrown during cleanup.", e.getMessage());
-        }
+        s3MockApi.stop();
+        FileSystemUtils.deleteRecursively(Paths.get(ROOT_DIR).toFile());
     }
 
     /**
-     * Test that the file is stored and the relative path is returned. If the file didn't exist then
-     * its relative path would not be found.
+     * Test that the file is stored in the S3 mock.
      */
     @Test
-    public void storeFileThatExists() {
+    void storeFileToS3ThatExists() {
         try {
             StorageFile storageFile = fileStorageService.storeFile(new MockMultipartFile("test", "test.txt",
-                    MediaType.TEXT_PLAIN_VALUE, "Test Pass-core".getBytes()));
+                    MediaType.TEXT_PLAIN_VALUE, "Test S3 Pass-core".getBytes()));
             assertFalse(fileStorageService.getResourceFileRelativePath(storageFile.getId()).isEmpty());
         } catch (Exception e) {
             assertEquals("An exception was thrown in storeFileThatExists.", e.getMessage());
@@ -73,13 +72,13 @@ public class FileStorageServiceTest {
     }
 
     /**
-     * File is stored and then retrieved.
+     * Should get the file from the S3 bucket and return it.
      */
     @Test
-    void getFileShouldReturnFile() {
+    void getFileFromS3ShouldReturnFile() {
         try {
             StorageFile storageFile = fileStorageService.storeFile(new MockMultipartFile("test", "test.txt",
-                    MediaType.TEXT_PLAIN_VALUE, "Test Pass-core".getBytes()));
+                    MediaType.TEXT_PLAIN_VALUE, "Test S3 Pass-core".getBytes()));
             ByteArrayResource file = fileStorageService.getFile(storageFile.getId());
             assertTrue(file.contentLength() > 0);
         } catch (IOException e) {
@@ -88,14 +87,12 @@ public class FileStorageServiceTest {
     }
 
     /**
-     * Should throw exception because file ID does not exist
+     * Should throw an exception because the file ID does not exist.
      */
     @Test
     void getFileShouldThrowException() {
         Exception exception = assertThrows(IOException.class,
-                () -> {
-                    ByteArrayResource file = fileStorageService.getFile("12345");
-                }
+                () -> fileStorageService.getFile("12345")
         );
         String expectedExceptionText = "File Service: The file could not be loaded";
         String actualExceptionText = exception.getMessage();
@@ -104,6 +101,7 @@ public class FileStorageServiceTest {
 
     @Test
     void storeFileWithDifferentLangFilesNames() {
+        //generate long string of all english characters
         String engFileName = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ!@#$%^&*()_+{}|:\"<>?`~[]\\;',./.txt";
         String frFileName = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZÀÁÂÃÄÅÇÈÉÊËÌÍÎÏÐÑÒÓÔÕÖÙÚÛÜÝßàáâãäå" +
                 "çèéêëìíîïñòóôõöùúûüýÿœŒæÆ.txt";
@@ -123,6 +121,7 @@ public class FileStorageServiceTest {
         allCharSets.put("ch", chFileName);
         allCharSets.put("ru", ruFileName);
 
+        //test each character set
         allCharSets.forEach((k,v) -> {
             try {
                 StorageFile storageFile = fileStorageService.storeFile(new MockMultipartFile("test", v,
@@ -136,7 +135,7 @@ public class FileStorageServiceTest {
     }
 
     /**
-     * Store file, then delete it. Should throw exception because the file was deleted.
+     * Stores file, then deletes it. Should throw an exception because the file does not exist.
      */
     @Test
     void deleteShouldThrowExceptionFileNotExist() {
@@ -145,14 +144,11 @@ public class FileStorageServiceTest {
                     MediaType.TEXT_PLAIN_VALUE, "Test Pass-core".getBytes()));
             fileStorageService.deleteFile(storageFile.getId());
             Exception exception = assertThrows(NotFoundException.class,
-                    () -> {
-                        fileStorageService.getResourceFileRelativePath(storageFile.getId());
-                    });
+                    () -> fileStorageService.getResourceFileRelativePath(storageFile.getId()));
             String exceptionText = exception.getMessage();
             assertTrue(exceptionText.matches("(.)+(was not found){1}(.)+"));
         } catch (IOException e) {
             assertEquals("Exception during deleteShouldThrowExceptionFileNotExist", e.getMessage());
         }
     }
-
 }
